Fixed comments + rebase
